requirements.txt
# NOTE: if TA-Lib got error, install this below for macOS
#1) brew install ta-lib
#2) xcode-select --install
#3) pip3 install TA-Lib

# Ubuntu - ARM (e.g. Odroid N2)
# wget -c http://prdownloads.sourceforge.net/ta-lib/ta-lib-0.4.0-src.tar.gz
# tar xvzf ta-lib-0.4.0-src.tar.gz
# cd ta-lib
# wget 'http://git.savannah.gnu.org/gitweb/?p=config.git;a=blob_plain;f=config.guess;hb=HEAD' -O config.guess
# wget 'http://git.savannah.gnu.org/gitweb/?p=config.git;a=blob_plain;f=config.sub;hb=HEAD' -O config.sub
# make
# sudo make install
# cd .. && pip3 install -r requirements.txt


# Installing below is by running pip3 install -r requirements.txt
python-binance
TA-Lib
numpy
websocket_client
redis
